refactor(demo): replace AppContext with useOneSignal hook
